% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESEQ_helpers.R
\name{annotate_gene_ensembl}
\alias{annotate_gene_ensembl}
\title{Collects human equivalent gene names and descriptions}
\usage{
annotate_gene_ensembl(data, organism = "oaries")
}
\arguments{
\item{data}{dataframe. Should be counts matrix with first
column name as 'gene_ensembl'}

\item{organism}{string. From gprofiler2: Organism names are constructed by
concatenating the first letter of the name and the family. For sheep data:
"oaries". For cattle data:"btaurus". Default is "oaries".}
}
\description{
\code{annotate_gene_ensembl} creates a dataframe of human equivalent gene
names and description souced from gprofiler2::gconvert.
}
\details{
The dataframe produced by this is used for downstream annotation of files.
It is also attached to the data when it is in the SummarizedExperiment class.
}
