% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_plot_multi_MA.R
\name{auto_plot_multi_MA}
\alias{auto_plot_multi_MA}
\title{Plot all MA plots automatically}
\usage{
auto_plot_multi_MA(
  PIF_list,
  list_plot_titles = NA,
  export_TIFFs = FALSE,
  export_pdf = FALSE,
  pdf_height = 140,
  pdf_width = 220,
  export_dir = "./outputs/MA_plots/",
  ...
)
}
\arguments{
\item{PIF_list}{List of dataframe produced by a call to
\code{auto_generate_DE_results()}}

\item{list_plot_titles}{A character vector or list of names in the same order
as \code{names(PIF_list)}. If NA (default) names are automatically
generated from  \code{names(PIF_list)}.}

\item{export_TIFFs}{Logical. Should individual TIFF images be exported for
each MA plot?}

\item{export_pdf}{Logical. Should all plots be exported to one pdf file?}

\item{pdf_height}{number. Height of each pdf page in mm.}

\item{pdf_width}{number. Width of each pdf page in mm.}

\item{export_dir}{string. Path directory where TIFF and PDF are exported to,
if exported.}

\item{...}{Other arguments parsed to \code{plot_custom_MA()}. \cr See
\code{?plot_custom_MA} for details of available arguments.}
}
\value{
Returns a named list of ggplot objects. Also exports plots to .TIFF
or .PDF, if requested.
}
\description{
Plot all MA plots automatically by iterating over a list of PIF dataframes,
naming each plot
}
\details{
Wrapper for \code{plot_custom_MA()} that uses sensible defaults and naming,
but is customisable.
}
