% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PIF.R
\name{calculate_PIF}
\alias{calculate_PIF}
\title{Calculate PIF}
\usage{
calculate_PIF(
  DE_res_list,
  top_level_groups = "LIV",
  log2_norm_data,
  colData,
  results_contrast_factor,
  export_tables = FALSE,
  export_dir = "./outputs/"
)
}
\arguments{
\item{DE_res_list}{List of DE results objects from DESEQ2. List names must be
in the format Condition 1 vs Condition 2 e.g. "LIV_HCP.HP.UMEI vs
LIV_LCP.LP.UMEI".}

\item{top_level_groups}{string to filter top level (e.g. Tissue Region).
Inherits from \code{auto_generate_DE_results()}}

\item{log2_norm_data}{log2 Normalised data from current call to DESEQ2 (NOT normalised data from WHOLE dataset)}

\item{colData}{coldata normally accessed via a call to \code{SummarizedExperiment::colData()}}

\item{results_contrast_factor}{non-string. See \code{?auto_generate_DE_results}}

\item{export_tables}{logical. Should an excel file be exported of results? Inherits from \code{auto_generate_DE_results()}}

\item{export_dir}{string. "./" indicates relative to working directory. IF
this directory doesn't exist, it will be created. Inherits from \code{auto_generate_DE_results()}.}
}
\value{
returns a list of dataframes with calculated mean expression, log2 FC and PIF columns.
}
\description{
Calculates Phenotypic Impact Factor (PIF) for each pairwise comparison.
}
\details{
The PIF value is the mean expression of Condition 1 and Condition 2 of the
pairwise comparison multiplied by the Log2 Fold Change. This favours genes
that are both highly abundant and highly different between the two
conditions. Designed to be used within the call to
\code{auto_generate_DE_results}, but can be used independently.
}
