% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RIF.R
\name{calculate_RIF}
\alias{calculate_RIF}
\title{Calculate RIF scores}
\usage{
calculate_RIF(
  DE_output,
  TFs = NULL,
  norm_exp_data,
  gene_annotations,
  all_genes_as_TF = FALSE,
  colData,
  results_contrast_factor,
  samples_colname,
  export_tables = FALSE,
  export_dir = "./outputs/"
)
}
\arguments{
\item{DE_output}{object (class = list) returned from a call to
\link{auto_generate_DE_results}}

\item{TFs}{character vector of Transcription Factors to use for RIF
calculations. Optional and only used if \code{all_genes_as_TF = FALSE}}

\item{norm_exp_data}{Whole data normalisation output - use VST norm.}

\item{gene_annotations}{Output from call to \link{annotate_gene_ensembl}}

\item{all_genes_as_TF}{Logical. Should all genes (with row mean expression >
2) be used as possible TFs in RIF?}

\item{colData}{coldata normally accessed via a call to
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::colData}}

\item{results_contrast_factor}{Name of Factor from
\code{DESeq2_formula_design} that is used to generate pairwise
comparisons.e.g. Region_Diet}

\item{samples_colname}{non-string. Name of \code{colData} column with
sample names (used for filtering).}

\item{export_tables}{logical. Should an excel file be exported of results? Default = FALSE}

\item{export_dir}{string. "./" indicates relative to working directory. IF
this directory doesn't exist, it will be created.}
}
\value{
Returns a list of dataframes, with each pairwise comparison having
own set of RIF scores: RIF1, RIF2 and mean_RIF (mean of RIF1 and RIF2, for
each gene)
}
\description{
Calculates 2 types of Regulatory Impact Factors, based on Reverter work.
Wrapper function for \link[CeTF:RIF]{CeTF::RIF}.
}
\details{
RIF scores are ... Genes do not need to be DE to be able to have a regulatory
effect...(hence, all_genes_as_TF)
}
