% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coldata}
\alias{coldata}
\title{coldata for testing}
\format{
A data frame with 239 rows and 11 variables:
\describe{
\item{sample_names}{Names of samples in correct format}
\item{Region_Diet}{Treatment information}
...
}
}
\source{
private
}
\usage{
coldata
}
\description{
A dataset of coldata from sheep experiment
}
\keyword{datasets}
