% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cts_all}
\alias{cts_all}
\title{Counts matrix for testing}
\format{
A data frame with 27054 rows and 240 variables:
\describe{
\item{gene_ensembl}{Ensembl IDs}
\item{Samples}{Each column is an individual sample, names match sample_names of coldata}
...
}
}
\source{
private
}
\usage{
cts_all
}
\description{
A counts matrix from sheep experiment, generated via feature counts.
}
\keyword{datasets}
