% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE4Rumi-package.R
\docType{package}
\name{DE4Rumi-package}
\alias{DE4Rumi}
\alias{DE4Rumi-package}
\title{DE4Rumi: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
