% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESEQ_helpers.R
\name{.calculate_DESEQ_results}
\alias{.calculate_DESEQ_results}
\title{Calculate DESEQ results}
\usage{
.calculate_DESEQ_results(
  dds_object,
  contrast_factor,
  combinations = NA,
  alpha = 0.05,
  lfc_thresh = 0,
  use_IHW_filtering
)
}
\arguments{
\item{dds_object}{The dds object, after DESeq2 has been run (e.g. dds_wald).
Normally with only 1 top-level filter (e.g. Tissue Region)}

\item{contrast_factor}{Non-string. Name of the column containing the factor
that will be contrasted. Should match one of the factors in DESeq2 design
(e.g. Region_Diet)}

\item{combinations}{Either NA (default) or a list with each element 2 strings
(e.g. "LIV_HCP-HP-UMEI" "LIV_LCP-LP-UMEI")}

\item{alpha}{numeric. What is the p-value threshold to be used for
determining significance. Used in call to \link[DESeq2:results]{DESeq2::results} and
others.}

\item{lfc_thresh}{numeric. What log2 Fold Change threshold should be applied
to determine significance in call to Used in call to \link[DESeq2:results]{DESeq2::results}.
Default is 0.}

\item{use_IHW_filtering}{Logical. Inherits from larger function. If \code{TRUE}
it will use a call to \code{\link[IHW:ihw.default]{IHW::ihw()}} when making results.}
}
\value{
Returns a named list with each element the results object output from
call to \link[DESeq2:results]{DESeq2::results}
}
\description{
Internal function to generate a list of DESEQ2 result objects for multiple
pairwise comparisons
}
\details{
This takes a dds object, which is pre-filtered to only contain one top-level
(e.g. Tissue Region) and uses the \link{make_pairwise_combinations} function
to generate all possible pairwise combinations, then use this to make all res
objects. It is designed to be used within larger function to automate DESeq2.
}
