% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{.GET_table_generic}
\alias{.GET_table_generic}
\title{GET tables}
\usage{
.GET_table_generic(map_string, DE_output, search_top_level = NULL)
}
\arguments{
\item{map_string}{character string that matches element names in the
auto_DE_output. Normally this is set by other functions.}

\item{DE_output}{object (class = list) returned from a call to
\link{auto_generate_DE_results}}

\item{search_top_level}{character vector of top_level_groups (see
\link{auto_generate_DE_results}) to search for, returning only tables from these
groups. e.g. \code{c("ARC","LIV")}. If not provided it will return all DE
tables.}
}
\value{
Returns a list of 1 level with all DE tables.
}
\description{
Internal accessor function that maps over the DE list output from
\link{auto_generate_DE_results}. Used by \link{GET_PIF_df}, \link{GET_DE_by_PIF_df}
and others.
}
