% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCIT.R
\name{.prep_PCIT_DE}
\alias{.prep_PCIT_DE}
\title{Prepare PIF data for PCIT}
\usage{
.prep_PCIT_DE(DE_full_out, search_top_level = NULL, DEn)
}
\arguments{
\item{DE_full_out}{Full output of \link{auto_generate_DE_results}}

\item{search_top_level}{character vector of top_level_groups (see
\link{auto_generate_DE_results}) to search for, returning only tables from these
groups. e.g. \code{c("ARC","LIV")}. If not provided it will return all DE
tables.}

\item{DEn}{Number of values to return, ranked on adjusted p value. E.g.
if \code{DEn = 20}, then the 20 genes with the lowest p value are returned
(there may be less than this due to prior filtering for significance).}
}
\value{
Returns a dataframe of all selected DE genes,
with summary columns showing why they were selected.
}
\description{
Internal function to prepare PIF data for PCIT
}
