% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCIT.R
\name{.prep_PCIT_PIF}
\alias{.prep_PCIT_PIF}
\title{Prepare PIF data for PCIT}
\usage{
.prep_PCIT_PIF(DE_full_out, search_top_level = NULL, PIF_sig, PIFn)
}
\arguments{
\item{DE_full_out}{Full output of \link{auto_generate_DE_results}}

\item{search_top_level}{character vector of top_level_groups (see
\link{auto_generate_DE_results}) to search for, returning only tables from these
groups. e.g. \code{c("ARC","LIV")}. If not provided it will return all DE
tables.}

\item{PIF_sig}{Number. Threshold used for filtering the z-scored PIF values, so a
\code{PIF_sig} of 2.58 filters for a nominal P value of 0.01, whereas a
value of 1.96 filters for a nominal p value of 0.05.}

\item{PIFn}{Number of values to return for both the max PIF and min PIF. E.g.
if \code{PIFn = 20}, then the 20 highest and 20 lowest are returned (if that many are present after filtering by PIF_sig)}
}
\value{
Returns a dataframe of all selected PIF genes,
with summary columns showing why they were selected.
}
\description{
Internal function to prepare PIF data for PCIT
}
