% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{.search_list_helper}
\alias{.search_list_helper}
\title{Helper to search levels for top_level_group}
\usage{
.search_list_helper(
  list_of_objects,
  search_term,
  pattern_extract_top_level = "^[^_]*"
)
}
\arguments{
\item{list_of_objects}{List of objects that's named. This is what is
searched. Inherited from other functions.}

\item{search_term}{character vector of top_level_groups (see
\link{auto_generate_DE_results}) to search for, returning only list entries from
these groups. e.g. \code{c("ARC","LIV")}. Inherited from other functions.}

\item{pattern_extract_top_level}{string representing a regex match to return
the names of the top level. Default works for all accessors from DE output.
RIF requires unique.}
}
\value{
Returns a filtered list.
}
\description{
Internal Helper to search levels for top_level_group. Used by
\link{.GET_table_generic}, \link{GET_pairwise_plots}
}
