% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_DE.R
\name{enrich_DE}
\alias{enrich_DE}
\title{Functional Enrichment of DE lists}
\usage{
enrich_DE(
  list_DE_obj,
  sig_thresh = 0.05,
  log2_thresh = 0,
  split_by_pos_neg = FALSE,
  base_URL = "http://biit.cs.ut.ee/gprofiler_archive3/e100_eg47_p14"
)
}
\arguments{
\item{list_DE_obj}{list of DESeq2 results objects, normally accessed via a
call to \code{auto_generate_DE_results()}.}

\item{sig_thresh}{number. Threshold to filter results by adjusted P value.
Defaults to 0.05.}

\item{log2_thresh}{number. Optional filter - if there is too many DE genes,
then list is filtered by log2 fold change. Will keep only DE genes with a FC
above this threshold. Uses absolute values (includes + and -).}

\item{split_by_pos_neg}{logical. Should the query be run separately for genes
with a + fold change, compared to genes with a - fold change?}

\item{base_URL}{string. URL for the version of g:Profiler to use. This
defaults to e100 as it matches the galaxy workflow, but also e103 broke a
lot of names and is yet unresolved. Provided in the format:
"http://biit.cs.ut.ee/gprofiler_archive3/e100_eg47_p14". \cr\cr Note that it
is "http://" NOT "https://". \cr\cr Providing NA will default to g:Profiler2
default of "http://biit.cs.ut.ee/gprofiler". \cr\cr All available archives
at: https://biit.cs.ut.ee/gprofiler/page/archives}
}
\value{
Returns a list of tables of same length as \code{list_DE_obj}, each with
g:Profiler results
}
\description{
Identify GO, REACTOME and KEGG pathways for DE lists via gProfiler's API
}
\details{
\code{enrich_DE()} takes the DE tables from a call to
\code{auto_generate_DE_results} as input to a call to
\code{gprofiler2::gost()}. Various custom options include: \itemize{\item
Filtering by adjusted p value or log2 fold change \item Split the DE list into
positive and negative fold change, to enrich separately. \item Remove parent terms from list
to de-clutter redudant terms. }
}
