% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{GET_boxplot_cooks_distance}
\alias{GET_boxplot_cooks_distance}
\title{GET boxplot of cooks distances}
\usage{
GET_boxplot_cooks_distance(auto_DE_output)
}
\arguments{
\item{auto_DE_output}{object (class = list) returned from a call to
\link{auto_generate_DE_results}}
}
\value{
Returns a boxplot.
}
\description{
Accessor function to access the boxplot of cooks distances for each sample,
generated by \link{auto_generate_DE_results}
}
