% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{GET_dds_wald_objects}
\alias{GET_dds_wald_objects}
\title{GET dds_wald objects}
\usage{
GET_dds_wald_objects(auto_DE_output, search_top_level = NULL)
}
\arguments{
\item{auto_DE_output}{object (class = list) returned from a call to
\link{auto_generate_DE_results}}

\item{search_top_level}{character vector of top_level_groups (see
\link{auto_generate_DE_results}) to search for, returning only tables from these
groups. e.g. \code{c("ARC","LIV")}. If not provided it will return all objects.}
}
\value{
Returns a list of 1 level with all dds_wald objects
(of class DESeqDataSet)
}
\description{
Accessor function to the DESeqDataSet objects (derived from the
SummarizedExperiment object parsed to \link{auto_generate_DE_results}) after the
call to DESeq2 (via \link{auto_generate_DE_results}).
}
