% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{GET_DE_by_PIF_df}
\alias{GET_DE_by_PIF_df}
\title{GET DE tables}
\usage{
GET_DE_by_PIF_df(auto_DE_output, search_top_level)
}
\arguments{
\item{auto_DE_output}{object (class = list) returned from a call to
\link{auto_generate_DE_results}}

\item{search_top_level}{character vector of top_level_groups (see
\link{auto_generage_DE_results})to search for, returning only DE tables of these
groups. If not provided it will return all DE tables. e.g. \code{c("ARC",
  "LIV")}}
}
\value{
Returns a list of 1 level with all DE tables.
}
\description{
Accessor function to access DE tables, sorted by PIF value, generated by
\link{auto_generate_DE_results}
}
