% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{GET_DESeq2_res_object}
\alias{GET_DESeq2_res_object}
\title{GET DESeq Results objects}
\usage{
GET_DESeq2_res_object(auto_DE_output, search_top_level = NULL)
}
\arguments{
\item{auto_DE_output}{object (class = list) returned from a call to
\link{auto_generate_DE_results}}

\item{search_top_level}{character vector of top_level_groups (see
\link{auto_generate_DE_results}) to search for, returning only tables from these
groups. e.g. \code{c("ARC","LIV")}. If not provided it will return all objects.}
}
\value{
Returns a list of 1 level with all DESeqResults objects
}
\description{
Accessor function to the results objects, containing all metadata, produced for
each pairwise comparison. Most useful for accessing advanced information on how
results were generated (via \link{auto_generate_DE_results}). This output is also
used by \link{enrich_DE}.
}
