% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{GET_normalised_data}
\alias{GET_normalised_data}
\title{GET normalised counts tables}
\usage{
GET_normalised_data(
  auto_DE_output,
  search_top_level = NULL,
  vst_or_log2 = c("VST", "Log2", "both")
)
}
\arguments{
\item{auto_DE_output}{object (class = list) returned from a call to
\link{auto_generate_DE_results}}

\item{search_top_level}{character vector of top_level_groups (see
\link{auto_generate_DE_results}) to search for, returning only tables from these
groups. e.g. \code{c("ARC","LIV")}. If not provided it will return all DE
tables.}

\item{vst_or_log2}{character vector of either: "VST", "Log2", "both"}
}
\value{
Returns a list of dataframes
}
\description{
Accessor function to access normalised counts tables for each top_level_group
individually, as generated by \link{auto_generate_DE_results}. These are generally
not useful, and downstream applications of RIF and PCIT, etc should use
the counts normalised across whole dataset.
}
