% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{GET_PIF_df}
\alias{GET_PIF_df}
\title{GET PIF tables}
\usage{
GET_PIF_df(auto_DE_output, search_top_level = NULL)
}
\arguments{
\item{auto_DE_output}{object (class = list) returned from a call to
\link{auto_generate_DE_results}}

\item{search_top_level}{character vector of top_level_groups (see
\link{auto_generate_DE_results}) to search for, returning only tables from these
groups. e.g. \code{c("ARC","LIV")}. If not provided it will return all DE
tables.}
}
\value{
Returns a list of 1 level with all DE tables.
}
\description{
Accessor function to access all genes with DE and PIF values, generated by
\link{auto_generate_DE_results}. Useful to prepare data for input to
\link{auto_plot_multi_MA} or \link{plot_custom_MA}.
}
