% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCIT.R
\name{PCIT_filter_output}
\alias{PCIT_filter_output}
\title{Filter PCIT output for Cytoscape}
\usage{
PCIT_filter_output(PCIT_out_data, cor_threshold, export_dir)
}
\arguments{
\item{PCIT_out_data}{Full output of \link{PCIT_calculate}. A list of 3 elements: tab, adj_raw, adj_sig.}

\item{cor_threshold}{Threshold for filtering PCIT correlations (both positive and negative)}

\item{export_dir}{}
}
\value{
Returns a ggplot/ggarrange object of density plots based on input
PCIT data. Also exports the required data for use in Cytoscape to
\code{export_dir}.
}
\description{
Function that takes the output from PCIT analysis and filters out insignificant
correlations, and further filters based on a hard threshold to allow plotting
only the most meaningful correlations (both positive and negative) in Cytoscape.
}
