% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_custom_MA.R
\name{plot_custom_MA}
\alias{plot_custom_MA}
\title{MA Plots}
\usage{
plot_custom_MA(
  PIF_data,
  padj_DE_filter = 0.05,
  PIF_sig = 2.58,
  n_labels_top_bottom = 20,
  label_type = "PIF & DE",
  custom_labels = NA,
  point_size = 1.5,
  point_alpha = 0.5,
  font_face = "plain",
  font_colour = "black",
  font_size = 8,
  label.rectangle = FALSE,
  plot_title = "TITLE",
  xlab = "Mean expression",
  ylab = "Log2 fold change",
  ggtheme = ggplot2::theme_classic(),
  palette = c("lightgrey", "orange", "red", "blue", "green"),
  export_TIFFs = FALSE,
  export_dir = "./outputs/MA_plots/",
  tiff_width = 190,
  tiff_height = 140,
  FC_thresh = 8
)
}
\arguments{
\item{PIF_data}{dataframe with at least the following columns: starts_with...
"gene_ensembl", "mean_", "log2FC_" and "zPIF"}

\item{padj_DE_filter}{Number. Significance threshold for determining if
points are labelled as DE or not.}

\item{PIF_sig}{Number. Used for filtering the z-scored PIF values, so a
\code{PIF_sig} of 2.58 filters for a nominal P value of 0.01, whereas a
value of 1.96 filters for a nominal p value of 0.05.}

\item{n_labels_top_bottom}{Number. How many labels should be displayed on
both the top and bottom. E.g. if n = 20, then 20 labels will be plotted
above y = 0 and 20 below y = 0.}

\item{label_type}{a string of either "PIF & DE" or "PIF". If "PIF & DE" is
selected, labels will only plot points that are BOTH PIF and DE, whereas
"PIF" will allow any point that is significantly PIF to be labelled.}

\item{custom_labels}{character vector of gene names to force labels for e.g.
c("CARTPT")}

\item{point_size}{Size of points on plot}

\item{point_alpha}{Transparency of points.}

\item{font_face}{String. Either "plain", "bold", "italic" or "bold.italic"}

\item{font_colour}{String of a colour name.}

\item{font_size}{Number. Size of font for labels.}

\item{label.rectangle}{logical. Should labels have a rectangle with white
background? See difference between \code{geom_label()} and
\code{geom_text()} by viewing \code{?ggplot2::geom_label}}

\item{plot_title}{String. Title of the plot.}

\item{xlab}{String. Title of x axis.}

\item{ylab}{String. Title of y axis.}

\item{ggtheme}{A function to a ggplot2 theme. E.g.
\code{ggplot2::theme_classic()}}

\item{palette}{character vector of colour names. Will label points in same
order as: \code{c("Not PIF or DE", "PIF", "PIF & DE", "DE", "Removed by
  DESeq2")}}

\item{export_TIFFs}{Logical. Should a TIFF image be exported?}

\item{export_dir}{string. "./" indicates relative to working directory. IF
this directory doesn't exist, it will be created.}

\item{tiff_width}{number in mm. Width of .tiff image if exported.}

\item{tiff_height}{number in mm. Height of .tiff image if exported.}

\item{FC_thresh}{Threshold for |Log2FC| for y-axis, default = 8.}
}
\value{
returns a ggplot object
}
\description{
Create labelled MA plots from previously calculated DE and PIF data.
}
\details{
These MA plots show the mean expression between Condition 1 and Condition 2
along the x axis, and the log2 fold change on the y axis. \cr Points that are
both highly abundant and highly different (higher/lower log2 fold change) are
considered to be most likely to be biologically relevant, and are represented
by the PIF calculation (see \code{calculate_PIF()}). \cr The points represent
the averages across multiple individual samples/animals. Therefore, points
that appear highly different may not be statistically different due to high
variance between the underlying points. Therefore, points that are both
Differentially Expressed (statistically different) and PIF are likely the
most important points. \cr Labels are generated automatically, and avoid too
many overlaps, see \code{?ggrepel::ggrepel}. Increasing the plot size will
increase number of labels. In RStudio, this is as simple as clicking on Zoom
in the 'Plots' window, or changing the \code{fig.width} and/or
\code{fig.length} parameters in a RMarkdown/RNotebook code chunk. \cr\cr
These plots are built with ggplot2. See https://ggplot2.tidyverse.org/ for
more details on the specs used.
}
