% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESEQ_helpers.R
\name{check_count_matrix}
\alias{check_count_matrix}
\title{Check formatting of supplied count matrix}
\usage{
check_count_matrix(data = cts)
}
\description{
\code{check_count_matrix()} returns messages to the console to determine if count
matrix has been formatted correctly for use in downstream analysis
}
\details{
The count matrix, normally as outputted from featureCounts, needs to be
formatted with:\itemize{
\item Column headings that match metadata
\item No rownames
\item First column titled gene_ensembl and contains all gene ensembl names
} This also provides a print out of the dimensions and predicted treatments
}
