% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE4Rumi-package.R
\docType{package}
\name{DE4Rumi-package}
\alias{DE4Rumi}
\alias{DE4Rumi-package}
\title{DE4Rumi: Differential Expression Analysis of RNAseq in Ruminant Models}
\description{
Wrappers for DESEQ2, custom plots and co-expression for a specific use-case related to ruminant RNSeq.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davos-i/DE4Rumi}
}

}
\author{
\strong{Maintainer}: David Innes \email{dave.j.innes@gmail.com} (\href{https://orcid.org/0000-0002-6246-0489}{ORCID})

}
\keyword{internal}
