% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_and_plot.R
\name{.norm_and_plot}
\alias{.norm_and_plot}
\title{Generate normalised counts and QC plots}
\usage{
.norm_and_plot(
  top_level_filter,
  dds_object,
  gene_annotations,
  export_tables = FALSE,
  export_dir = "./outputs/normalised_counts/",
  results_contrast_factor
)
}
\arguments{
\item{top_level_filter}{inherits from \code{auto_generate_DE_results()}}

\item{dds_object}{DESeq2 data object, inherits from
\code{auto_generate_DE_results()}}

\item{gene_annotations}{Output from \code{annotate_gene_ensembl()}}

\item{export_tables}{logical. \code{TRUE} indicates the normalised counts
tables, both vsd and log2norm, annotated with gene names and descriptions,
are exported to \code{"./outputs/normalised_counts/"})}

\item{export_dir}{string. "./" indicates relative to working directory. IF
this directory doesn't exist, it will be created.}

\item{results_contrast_factor}{inherits from
\code{auto_generate_DE_results()}}
}
\value{
returns a list of 1) heatmap_plot, PCA_plot and Sample distances
heatmap; and 2) normalised vsd and normalised log2norm dataframes.
}
\description{
Internal function to use within \code{auto_generate_DE_results()}.
}
\details{
This function generates normalised count tables, using the variance
stabilizing transformation technique (called "vsd" throughougt - variance
stabilised data), and also using log2 normilisation (log2norm). This function
also generates 3 types of quality control plots for the overall dataset.
}
