% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{GET_pairwise_plots}
\alias{GET_pairwise_plots}
\title{GET pairwise plots}
\usage{
GET_pairwise_plots(
  auto_DE_output,
  type = c("PCA", "heatmap", "sample_distances")
)
}
\arguments{
\item{auto_DE_output}{object (class = list) returned from a call to
\link{auto_generate_DE_results}}

\item{type}{one of either: "PCA", "heatmap" or "sample_distances"}
}
\value{
Returns a list of 1 level with all plots of \code{type}. Also prints
plots, particularly useful in a RMarkdown code chunk.
}
\description{
Accessor function to access one of three types of pairwise plots generated
by \link{auto_generate_DE_results}
}
