% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCIT.R
\name{PCIT_calculate}
\alias{PCIT_calculate}
\title{Run PCIT calculation}
\usage{
PCIT_calculate(norm_counts_for_PCIT, key_colname = "gene_name")
}
\arguments{
\item{norm_counts_for_PCIT}{Normalised counts for selected genes. Normally
this is produced by a call to \link{PCIT_prepare_data}.}

\item{key_colname}{string of the name of column to be used as key, typically
\code{"gene_ensembl"} or \code{"gene_name"}.}
}
\value{
Returns a list of: the significant correlations, raw adjacency matrix and significant adjacency matrix.
}
\description{
Wrapper for \link[CeTF:PCIT]{CeTF::PCIT} which is based on Reverter & Chan 2008. This can take
a long time, especially for larger datasets.
}
